requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.10
cookiecutter==2.6.0
dearpygui==1.10.1
decorator==4.4.2
distro==1.9.0
fastapi==0.115.0
filelock==3.13.1
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
frozenlist==1.5.0
fsspec==2023.12.2
g4f==0.5.0.4
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.29.3
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
jiter==0.9.0
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mido==1.3.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.2.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.0
packaging==23.2
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pefile==2023.2.7
Pillow==10.0.1
pretty_midi==0.2.10
proglog==0.1.10
progress==1.6
propcache==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.22.0
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.5.1
Pygments==2.18.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyttsx3==2.98
pywin32==310
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
regex==2024.11.6
repath==0.9.0
requests==2.31.0
rich==13.9.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
sentence-transformers==3.4.1
shiboken6==6.6.0
silero==0.4.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
srt==3.5.3
starlette==0.38.6
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.1
transformers==4.49.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
ui==0.1.4
urllib3==2.0.5
ursina==7.0.0
uvicorn==0.31.1
vosk==0.3.45
watchdog==4.0.2
watchfiles==0.24.0
Wave==0.0.2
websockets==12.0
yarl==1.18.3
